Resources externalization, project restructuration
