+++ src/components/intent-components/intent-list/intent-list.styles.ts
