+++ src/components/intent-components/intent-list/package.json
