+++ src/components/intent-components/training-selection-options/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "training-selection-options.tsx"
+}
