+++ src/localization/en.json
@@ -32,13 +32,13 @@
     "storyEditor": "Story editor",
     "globalEditor": "Global editor",
     "leftBar" : {
+      "storyLeftTab": "Story",
       "knotsLeftTab": "Knots",
       "intentsLeftTab": "Intents",
       "knotSearchHelper": "Knot search",
       "intentSearchHelper": "Search for an intent"
     },
     "rightBar" : {
-      "storyRightTab": "Story",
       "detailsRightTab": "Details",
       "linkingRightTab": "Linking",
       "storyNameHelper": "Story name",
